update duration setting on time sensor to make the range more usable

+++ plugins/paintops/libpaintop/sensors/kis_dynamic_sensor_time.h
@@ -38,7 +38,7 @@ public:
     virtual QWidget* createConfigurationWidget(QWidget* parent, QWidget*);
 public Q_SLOTS:
     virtual void setPeriodic(bool periodic);
-    virtual void setLength(int length);
+    virtual void setLength(qreal length);
 
 
     virtual void toXML(QDomDocument&, QDomElement&) const;
